Rename request/response files -> requests/responses, update imports, update docs
