+++ mkdocs.yml
@@ -19,7 +19,7 @@ nav:
     - Applications: 'applications.md'
     - Test Client: 'test_client.md'
     - Debugging: 'debugging.md'
-    - Views: 'views.md'
+    - Endpoints: 'endpoints.md'
 
 markdown_extensions:
   - markdown.extensions.codehilite:
