+++ starlette/endpoints.py
@@ -1,6 +1,6 @@
 from starlette.exceptions import HTTPException
-from starlette.request import Request
-from starlette.response import Response, PlainTextResponse
+from starlette.requests import Request
+from starlette.responses import Response, PlainTextResponse
 from starlette.types import Receive, Send, Scope
 import asyncio
